auto-integrating nitro
