+++ src/commands/deploy/nextjs/deploy.ts
@@ -1,11 +1,7 @@
-import fs, { existsSync, readFileSync } from "fs";
+import fs, { existsSync } from "fs";
 import { GenezioDeployOptions } from "../../../models/commandOptions.js";
-import { YamlConfigurationIOController } from "../../../projectConfiguration/yaml/v2.js";
 import { YamlProjectConfiguration } from "../../../projectConfiguration/yaml/v2.js";
-import inquirer from "inquirer";
 import path from "path";
-import { regions } from "../../../utils/configs.js";
-import { checkProjectName } from "../../create/create.js";
 import { debugLogger, log } from "../../../utils/logging.js";
 import { $ } from "execa";
 import { UserError } from "../../../errors.js";
@@ -35,20 +31,14 @@ import {
     NEXT_JS_GET_ACCESS_KEY,
     NEXT_JS_GET_SECRET_ACCESS_KEY,
 } from "../../../constants.js";
-import getProjectInfoByName from "../../../requests/getProjectInfoByName.js";
 import colors from "colors";
-import {
-    uniqueNamesGenerator,
-    adjectives,
-    colors as ungColors,
-    animals,
-} from "unique-names-generator";
 import { getPresignedURLForProjectCodePush } from "../../../requests/getPresignedURLForProjectCodePush.js";
 import { computeAssetsPaths } from "./assets.js";
 import * as Sentry from "@sentry/node";
 import { randomUUID } from "crypto";
 import { EdgeFunction, getEdgeFunctions } from "./edge.js";
 import { uploadEnvVarsFromFile } from "../utils.js";
+import { readOrAskConfig } from "../utils.js";
 
 export async function nextJsDeploy(options: GenezioDeployOptions) {
     // Check if node_modules exists
@@ -500,75 +490,4 @@ async function writeOpenNextConfig(region: string, edgeFunctionPaths: EdgeFuncti
 
     const tag = ENVIRONMENT === "prod" ? "latest" : "dev";
     await getPackageManager().install([`@genezio/nextjs-isr-${region}@${tag}`]);
-}
-
-async function readOrAskConfig(configPath: string): Promise<YamlProjectConfiguration> {
-    const configIOController = new YamlConfigurationIOController(configPath);
-    if (!existsSync(configPath)) {
-        const name = await readOrAskProjectName();
-
-        let region = regions[0].value;
-        if (process.env["CI"] !== "true") {
-            ({ region } = await inquirer.prompt([
-                {
-                    type: "list",
-                    name: "region",
-                    message: "Select the Genezio project region:",
-                    choices: regions,
-                },
-            ]));
-        } else {
-            log.info(
-                "Using the default region for the project because no `genezio.yaml` file was found.",
-            );
-        }
-
-        await configIOController.write({ name, region, yamlVersion: 2 });
-    }
-
-    return await configIOController.read();
-}
-
-async function readOrAskProjectName(): Promise<string> {
-    if (existsSync("package.json")) {
-        // Read package.json content
-        const packageJson = readFileSync("package.json", "utf-8");
-        const packageJsonName = JSON.parse(packageJson)["name"];
-
-        const validProjectName: boolean = await (async () => checkProjectName(packageJsonName))()
-            .then(() => true)
-            .catch(() => false);
-
-        const projectExists = await getProjectInfoByName(packageJsonName)
-            .then(() => true)
-            .catch(() => false);
-
-        // We don't want to automatically use the package.json name if the project
-        // exists, because it could overwrite the existing project by accident.
-        if (packageJsonName !== undefined && validProjectName && !projectExists)
-            return packageJsonName;
-    }
-
-    let name = uniqueNamesGenerator({
-        dictionaries: [ungColors, adjectives, animals],
-        separator: "-",
-        style: "lowerCase",
-        length: 3,
-    });
-    if (process.env["CI"] !== "true") {
-        // Ask for project name
-        ({ name } = await inquirer.prompt([
-            {
-                type: "input",
-                name: "name",
-                message: "Enter the Genezio project name:",
-                default: path.basename(process.cwd()),
-                validate: checkProjectName,
-            },
-        ]));
-    } else {
-        log.info("Using a random name for the project because no `genezio.yaml` file was found.");
-    }
-
-    return name;
 }
