+++ src/commands/deploy/nitrojs/deploy.ts
@@ -1,34 +1,23 @@
 import { $ } from "execa";
 import { GenezioDeployOptions } from "../../../models/commandOptions.js";
 import { UserError } from "../../../errors.js";
-import { YamlConfigurationIOController } from "../../../projectConfiguration/yaml/v2.js";
 import { YamlProjectConfiguration } from "../../../projectConfiguration/yaml/v2.js";
-import { regions } from "../../../utils/configs.js";
-import inquirer from "inquirer";
-import { log } from "../../../utils/logging.js";
-import { existsSync, readFileSync } from "fs";
 import { functionToCloudInput, getCloudAdapter } from "../genezio.js";
-import { checkProjectName } from "../../create/create.js";
-import getProjectInfoByName from "../../../requests/getProjectInfoByName.js";
-import {
-    uniqueNamesGenerator,
-    adjectives,
-    colors as ungColors,
-    animals,
-} from "unique-names-generator";
-import path from "path";
 import { getCloudProvider } from "../../../requests/getCloudProvider.js";
 import { FunctionType, Language } from "../../../projectConfiguration/yaml/models.js";
 import { PackageManagerType } from "../../../packageManagers/packageManager.js";
 import { ProjectConfiguration } from "../../../models/projectConfiguration.js";
 import { debugLogger } from "../../../utils/logging.js";
+import { readOrAskConfig } from "../utils.js";
+import { existsSync } from "fs";
+import { getPackageManager } from "../../../packageManagers/packageManager.js";
 export async function nitroJsDeploy(options: GenezioDeployOptions) {
     // Check if node_modules exists
-    // if (!existsSync("node_modules")) {
-    //     throw new UserError(
-    //         `Please run \`${getPackageManager().command} install\` before deploying your Next.js project. This will install the necessary dependencies.`,
-    //     );
-    // }
+    if (!existsSync("node_modules")) {
+        throw new UserError(
+            `Please run \`${getPackageManager().command} install\` before deploying your Next.js project. This will install the necessary dependencies.`,
+        );
+    }
     await $({ stdio: "inherit" })`npx nitro build --preset aws_lambda`.catch(() => {
         throw new UserError("Failed to build the Nitro.js project. Check the logs above.");
     });
@@ -42,7 +31,7 @@ async function deployFunctions(config: YamlProjectConfiguration, stage?: string)
 
     const functions = [
         {
-            path: `server`,
+            path: `./.output/server`,
             name: "test-nitro",
             entry: "index.mjs",
             handler: "handler",
@@ -53,7 +42,7 @@ async function deployFunctions(config: YamlProjectConfiguration, stage?: string)
     const deployConfig: YamlProjectConfiguration = {
         ...config,
         backend: {
-            path: ".output/",
+            path: "",
             language: {
                 name: Language.js,
                 runtime: "nodejs20.x",
@@ -85,75 +74,4 @@ async function deployFunctions(config: YamlProjectConfiguration, stage?: string)
     debugLogger.debug(`Deployed functions: ${JSON.stringify(result.functions)}`);
 
     return result;
-}
-
-async function readOrAskConfig(configPath: string): Promise<YamlProjectConfiguration> {
-    const configIOController = new YamlConfigurationIOController(configPath);
-    if (!existsSync(configPath)) {
-        const name = await readOrAskProjectName();
-
-        let region = regions[0].value;
-        if (process.env["CI"] !== "true") {
-            ({ region } = await inquirer.prompt([
-                {
-                    type: "list",
-                    name: "region",
-                    message: "Select the Genezio project region:",
-                    choices: regions,
-                },
-            ]));
-        } else {
-            log.info(
-                "Using the default region for the project because no `genezio.yaml` file was found.",
-            );
-        }
-
-        await configIOController.write({ name, region, yamlVersion: 2 });
-    }
-
-    return await configIOController.read();
-}
-
-async function readOrAskProjectName(): Promise<string> {
-    if (existsSync("package.json")) {
-        // Read package.json content
-        const packageJson = readFileSync("package.json", "utf-8");
-        const packageJsonName = JSON.parse(packageJson)["name"];
-
-        const validProjectName: boolean = await (async () => checkProjectName(packageJsonName))()
-            .then(() => true)
-            .catch(() => false);
-
-        const projectExists = await getProjectInfoByName(packageJsonName)
-            .then(() => true)
-            .catch(() => false);
-
-        // We don't want to automatically use the package.json name if the project
-        // exists, because it could overwrite the existing project by accident.
-        if (packageJsonName !== undefined && validProjectName && !projectExists)
-            return packageJsonName;
-    }
-
-    let name = uniqueNamesGenerator({
-        dictionaries: [ungColors, adjectives, animals],
-        separator: "-",
-        style: "lowerCase",
-        length: 3,
-    });
-    if (process.env["CI"] !== "true") {
-        // Ask for project name
-        ({ name } = await inquirer.prompt([
-            {
-                type: "input",
-                name: "name",
-                message: "Enter the Genezio project name:",
-                default: path.basename(process.cwd()),
-                validate: checkProjectName,
-            },
-        ]));
-    } else {
-        log.info("Using a random name for the project because no `genezio.yaml` file was found.");
-    }
-
-    return name;
 }
